chore: Restructure and extract index functionality into modules for BPC-4.
