+++ package.json
@@ -6,7 +6,7 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node ."
+    "start": "node src/index.js"
   },
   "dependencies": {
     "@types/node": "^18.0.6",
