+++ src/handlers/end.js
@@ -1,0 +1,4 @@
+// end is called when your Battlesnake finishes a game
+export default function end(gameState) {
+    console.log("GAME OVER\n");
+  }
