+++ src/handlers/start.js
@@ -1,0 +1,4 @@
+// start is called when your Battlesnake begins a game
+export default function start(gameState) {
+    console.log("GAME START");
+  }
