+++ src/index.js
@@ -1,0 +1,24 @@
+// Welcome to
+// __________         __    __  .__                               __
+// \______   \_____ _/  |__/  |_|  |   ____   ______ ____ _____  |  | __ ____
+//  |    |  _/\__  \\   __\   __\  | _/ __ \ /  ___//    \\__  \ |  |/ // __ \
+//  |    |   \ / __ \|  |  |  | |  |_\  ___/ \___ \|   |  \/ __ \|    <\  ___/
+//  |________/(______/__|  |__| |____/\_____>______>___|__(______/__|__\\_____>
+//
+// This file can be a nice home for your Battlesnake logic and helper functions.
+//
+// To get you started we've included code to prevent your Battlesnake from moving backwards.
+// For more info see docs.battlesnake.com
+
+import runServer from './server.js';
+import info from './handlers/info.js';
+import start from './handlers/start.js';
+import move from './handlers/move.js';
+import end from './handlers/end.js';
+
+runServer({
+  info: info,
+  start: start,
+  move: move,
+  end: end
+});
